Move treemapLayout.js to more logical place
